// ui/lib/api/schemas/textData.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 */
import type { BlobRef } from './blobRef';
import type { FontPrediction } from './fontPrediction';
import type { TextDirection } from './textDirection';
import type { TextStyle } from './textStyle';
import type { TextWorkflow } from './textWorkflow';
import type { Transform } from './transform';

export interface TextData {
  confidence?: number;
  /** @nullable */
  detectedFontSizePx?: number | null;
  /** @nullable */
  detector?: string | null;
  fontPrediction?: null | FontPrediction;
  /** @nullable */
  linePolygons?: number[][][] | null;
  lockLayoutBox?: boolean;
  renderedDirection?: null | TextDirection;
  /** @nullable */
  rotationDeg?: number | null;
  sourceDirection?: null | TextDirection;
  /** @nullable */
  sourceLang?: string | null;
  sprite?: null | BlobRef;
  spriteTransform?: null | Transform;
  style?: null | TextStyle;
  /** @nullable */
  text?: string | null;
  /** @nullable */
  translation?: string | null;
  workflow?: TextWorkflow;
}